forbid scenario deletion during a run

+++ scenario/src/main/kotlin/com/cosmotech/scenario/service/ScenarioServiceImpl.kt
@@ -18,6 +18,7 @@ import com.cosmotech.api.events.ScenarioLastRunChanged
 import com.cosmotech.api.events.ScenarioRunEndToEndStateRequest
 import com.cosmotech.api.events.ScenarioRunStartedForScenario
 import com.cosmotech.api.events.WorkflowPhaseToStateRequest
+import com.cosmotech.api.exceptions.CsmClientException
 import com.cosmotech.api.exceptions.CsmResourceNotFoundException
 import com.cosmotech.api.rbac.CsmRbac
 import com.cosmotech.api.rbac.PERMISSION_CREATE_CHILDREN
@@ -236,6 +237,8 @@ internal class ScenarioServiceImpl(
     val scenario = this.findScenarioById(organizationId, workspaceId, scenarioId)
     csmRbac.verify(scenario.getRbac(), PERMISSION_DELETE, scenarioPermissions)
 
+    if (scenario.state == ScenarioJobState.Running)
+        throw CsmClientException("Can't delete a running scenario : ${scenario.id}")
     scenarioRepository.delete(scenario)
 
     this.handleScenarioDeletion(organizationId, workspaceId, scenario)
